requirements.txt
backports.functools-lru-cache==1.6.4
blis==0.7.9
Bottleneck==1.3.5
brotlipy==0.7.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chainer==7.8.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.6
confection==0.0.3
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.7
cytoolz==0.11.0
dask==2022.7.0
docopt==0.6.2
filelock==3.8.0
fsspec==2022.3.0
ftfy==6.1.1
greenlet==1.1.1
huggingface-hub==0.0.8
idna==3.4
imagecodecs==2021.8.26
imageio==2.19.3
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.2
langcodes==3.3.0
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.2.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
murmurhash==1.0.9
networkx==2.8.4
numexpr==2.8.3
numpy==1.22.3
overrides==7.3.1
packaging==21.3
pandas==1.4.3
partd==1.2.0
pathy==0.8.1
Pillow==9.2.0
pip==22.2.2
pipreqs==0.4.11
portalocker==2.6.0
preshed==3.0.8
protobuf==3.20.1
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacrebleu==1.4.10
sacremoses==0.0.53
scikit-image==0.19.2
scipy==1.7.3
setuptools==65.5.0
shellingham==1.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.39
srsly==2.4.5
thinc==8.0.17
tifffile==2021.7.2
timm==0.4.12
tokenizers==0.10.3
toolz==0.11.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.1
transformers==4.7.0
typer==0.4.2
typing_extensions==4.3.0
typing-utils==0.1.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
wheel==0.37.1
yarg==0.1.9
zipp==3.10.0
clip @ git+https://github.com/openai/CLIP.git
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl